# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arc LMS API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Arc LMS Support",
      "email": "support@arc-lms.com",
      "url": "https://arc-lms.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://arc-lms.onrender.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/meetings": {
      "get": {
        "tags": ["Meetings"],
        "summary": "List meetings",
        "operationId": "listMeetings",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "status", "in": "query", "schema": {"type": "string", "enum": ["SCHEDULED", "LIVE", "ENDED", "CANCELLED"]}}],
        "responses": {
          "200": {"description": "Meetings retrieved", "content": {"application/json": {"schema": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Meeting"}}}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "post": {
        "tags": ["Meetings"],
        "summary": "Schedule meeting",
        "operationId": "scheduleMeeting",
        "security": [{"BearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScheduleMeetingRequest"}}}},
        "responses": {
          "201": {"description": "Meeting scheduled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Meeting"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/meetings/{id}": {
      "get": {
        "tags": ["Meetings"],
        "summary": "Get meeting by ID",
        "operationId": "getMeeting",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Meeting retrieved", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Meeting"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "put": {
        "tags": ["Meetings"],
        "summary": "Update meeting",
        "operationId": "updateMeeting",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateMeetingRequest"}}}},
        "responses": {
          "200": {"description": "Meeting updated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Meeting"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/meetings/{id}/start": {
      "post": {
        "tags": ["Meetings"],
        "summary": "Start meeting",
        "operationId": "startMeeting",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "responses": {
          "200": {"description": "Meeting started", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Meeting"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/meetings/{id}/end": {
      "post": {
        "tags": ["Meetings"],
        "summary": "End meeting",
        "operationId": "endMeeting",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "responses": {
          "200": {"description": "Meeting ended", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Meeting"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/meetings/{id}/join": {
      "get": {
        "tags": ["Meetings"],
        "summary": "Get meeting join URL",
        "operationId": "joinMeeting",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Join URL retrieved", "content": {"application/json": {"schema": {"type": "object", "properties": {"join_url": {"type": "string", "format": "uri"}}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /public/auth/login"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "description": "Unique key to ensure idempotent requests. Prevents duplicate operations.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or token invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required",
                "details": {}
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "Insufficient permissions to perform this action",
                "details": {}
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "Resource not found",
                "details": {}
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Meeting": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string", "format": "uuid"},
          "title": {"type": "string"},
          "description": {"type": "string"},
          "host_id": {"type": "string", "format": "uuid"},
          "host": {"$ref": "#/components/schemas/User"},
          "class_id": {"type": "string", "format": "uuid"},
          "course_id": {"type": "string", "format": "uuid"},
          "scheduled_start": {"type": "string", "format": "date-time"},
          "scheduled_end": {"type": "string", "format": "date-time"},
          "actual_start": {"type": "string", "format": "date-time"},
          "actual_end": {"type": "string", "format": "date-time"},
          "status": {"type": "string", "enum": ["SCHEDULED", "LIVE", "ENDED", "CANCELLED"]},
          "meeting_url": {"type": "string", "format": "uri"},
          "recording_url": {"type": "string", "format": "uri"},
          "max_participants": {"type": "integer"},
          "participant_count": {"type": "integer"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "ScheduleMeetingRequest": {
        "type": "object",
        "required": ["title", "scheduled_start", "scheduled_end"],
        "properties": {
          "title": {"type": "string", "minLength": 1, "maxLength": 200},
          "description": {"type": "string", "maxLength": 1000},
          "class_id": {"type": "string", "format": "uuid"},
          "course_id": {"type": "string", "format": "uuid"},
          "scheduled_start": {"type": "string", "format": "date-time"},
          "scheduled_end": {"type": "string", "format": "date-time"},
          "max_participants": {"type": "integer", "minimum": 2}
        }
      },
      "UpdateMeetingRequest": {
        "type": "object",
        "properties": {
          "title": {"type": "string", "minLength": 1, "maxLength": 200},
          "description": {"type": "string", "maxLength": 1000},
          "scheduled_start": {"type": "string", "format": "date-time"},
          "scheduled_end": {"type": "string", "format": "date-time"},
          "max_participants": {"type": "integer", "minimum": 2}
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error context",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": ["id", "role", "email", "first_name", "last_name", "status", "created_at", "updated_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "NULL for SUPER_ADMIN"
          },
          "role": {
            "type": "string",
            "enum": ["SUPER_ADMIN", "ADMIN", "TUTOR", "STUDENT"]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "first_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "last_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "middle_name": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "profile_photo": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE", "DEACTIVATED", "PENDING"]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of resource:action permissions (e.g., 'tenant:create', 'user:delete')"
          },
          "notification_preferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationPreference"
            }
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivation_reason": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationPreference": {
        "type": "object",
        "required": ["event_type", "in_app_enabled", "push_enabled", "email_enabled"],
        "properties": {
          "event_type": {
            "type": "string"
          },
          "in_app_enabled": {
            "type": "boolean"
          },
          "push_enabled": {
            "type": "boolean"
          },
          "email_enabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}